backend/requirements.txt
mediapipe==0.10.9
opencv-python==4.8.1.78
numpy==1.24.3
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0 